requirements.txt
google-api-python-client
google-generativeai
markdown
playwright
requests
Pillow
